feat: agrego test y actualizo version de Spring en ms-productos-personalizados

+++ ms-productopersonalizado/src/main/java/com/frankester/msproductopersonalizado/models/Personalizacion.java
@@ -3,6 +3,8 @@ package com.frankester.msproductopersonalizado.models;
 import jakarta.persistence.*;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Entity
 public class Personalizacion {
@@ -11,7 +13,7 @@ public class Personalizacion {
     private Long id;
 
     private String nombre;
-    private Integer precio;
+    private BigDecimal precio;
     private String contenido;
     
     @ManyToOne(fetch = FetchType.EAGER, cascade = CascadeType.ALL)
