+++ ms-productopersonalizado/src/main/java/com/frankester/msproductopersonalizado/models/dto/ProductoPersonalizadoDTO.java
@@ -4,11 +4,13 @@ import com.frankester.msproductopersonalizado.models.PosiblePersonalizacion;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ProductoPersonalizadoDTO {
     private String productoBaseNombre;
     private String nombre;
-    private Integer precio;
+    private BigDecimal precio;
     private String contenido;
     private String vendedorCuil;
     private PosiblePersonalizacion posiblePersonalizacion;
